Changed behaviour of interactive funcs in vis.mpl
The functions where using pyplot.show() to stay open while the user
picks points, etc. This causes a problem when the plot is interactive
(in IPython with -pylab). In those cases, show doesn't hold.

The alternative is to return a function instead of the picked points,
etc. This function returns what has been picked when it is called. So in
interactive mode, after the picking is done, just call the function
returned to get the values.